backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
httpx==0.26.0
pyjwt==2.8.0
langgraph==0.0.48
langchain-google-genai==0.0.6
openai==1.12.0
